+++ dali/pipeline/operators/color/color_twist.cu
@@ -32,7 +32,7 @@ void ColorTwistBase<GPUBackend>::RunImpl(DeviceWorkspace *ws, const int idx) {
   cudaStream_t old_stream = nppGetStream();
   nppSetStream(ws->stream());
 
-  for (int i = 0; i < input.ntensor(); ++i) {
+  for (size_t i = 0; i < input.ntensor(); ++i) {
     if (!augments_.empty()) {
       float matrix[nDim][nDim];
       float * m = reinterpret_cast<float*>(matrix);
